Remove @Autowired annotation in constructors because:
As of Spring Framework 4.3, an @Autowired annotation on such a
constructor is no longer necessary if the target bean defines only one
constructor to begin with. However, if several constructors are
available, at least one must be annotated with @Autowired in order to
instruct the container which one to use.